Enable eval() to be Used With Function Arguments as Decorator
The eval() function is now updated to use the arguments of
the function as potential modules when used as a decorator.

The arguments passed to the decorated function will be checked
and only nn.Modules will be passed to the context manager.

This makes it unnecessary to have global modules.