Fix eval() backward() Issue
The eval() context manager was designed to put all modules to
eval mode and set gradients off in one shot.

However, while training on GPUs, one could not call backward()
for some strange reason.

It turned out that the eval() method reset the training attribute
of the modules.

This, however is only a boolean flag that should only be read
(to have different behaviours, for example. Like dropout).

One can see this peculiarity in action by calling mag.eval() on
a module and then printing [m.training for m in m.modules()].

The solution is, of course, to replace .training with .train().